fix: move did related utils to did pkg
renamed util.go to utils.go for consistency.  we generally use utils.go
as the name for these types of files.
